# benchmark and optimize httpevents