--- a/event_safe.go
+++ b/event_safe.go
@@ -9,3 +9,7 @@ func cloneValue(v interface{}) interface{} {
 func bytesToString(b []byte) string {
 	return string(b)
 }
+
+func noescape(a []interface{}) []interface{} {
+	return a
+}