--- a/event_safe.go
+++ b/event_safe.go
@@ -1,0 +1,11 @@
+// +build !go1.5
+
+package events
+
+func cloneValue(v interface{}) interface{} {
+	return v
+}
+
+func bytesToString(b []byte) string {
+	return string(b)
+}