--- a/event_unsafe.go
+++ b/event_unsafe.go
@@ -32,10 +32,10 @@ func bytesToString(b []byte) string {
 }
 
 func noescape(a []interface{}) (b []interface{}) {
-	// The use of unsafe.Pointer tricks the compiler into thinking that the
+	// The use of reflect.SliceHeader tricks the compiler into thinking that the
 	// content of the input slice doesn't escape, so we can prevent dynamic
-	// memory allocations that would otherwise happen for each argument of
-	// a log message.
+	// memory allocations that would otherwise happen for each argument of a log
+	// message.
 	*(*reflect.SliceHeader)(unsafe.Pointer(&b)) = *(*reflect.SliceHeader)(unsafe.Pointer(&a))
 	return
 }