--- a/httpevents/handler_test.go
+++ b/httpevents/handler_test.go
@@ -4,9 +4,7 @@ import (
 	"context"
 	"net/http"
 	"net/http/httptest"
-	"reflect"
 	"testing"
-	"time"
 
 	"github.com/segmentio/events"
 )
@@ -65,11 +63,7 @@ func TestHandler(t *testing.T) {
 		return
 	}
 
-	e := *evList[0]
-	e.Source = ""
-	e.Time = time.Time{}
-
-	if !reflect.DeepEqual(e, events.Event{
+	if e := *evList[0]; !equalEvents(e, events.Event{
 		Message: `127.0.0.1:80->127.0.0.1:56789 - www.github.com - GET /hello?answer=42#universe - 202 Accepted - "httpevents"`,
 		Args: events.Args{
 			{"local_address", "127.0.0.1:80"},
@@ -119,11 +113,7 @@ func TestHandlerPanic(t *testing.T) {
 		return
 	}
 
-	e := *evList[0]
-	e.Source = ""
-	e.Time = time.Time{}
-
-	if !reflect.DeepEqual(e, events.Event{
+	if e := *evList[0]; !equalEvents(e, events.Event{
 		Message: `127.0.0.1:80->127.0.0.1:56789 - www.github.com - POST / - 500 Internal Server Error - "httpevents"`,
 		Args: events.Args{
 			{"local_address", "127.0.0.1:80"},
@@ -138,6 +128,26 @@ func TestHandlerPanic(t *testing.T) {
 		t.Errorf("%#v", e)
 	}
 }
+
+func BenchmarkHandler(b *testing.B) {
+	l := &events.Logger{}
+	h := NewHandlerWith(l, http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
+		res.WriteHeader(http.StatusOK)
+	}))
+
+	w := &mockResponseWriter{}
+	r := httptest.NewRequest("GET", "http://localhost:4242/", nil)
+
+	for i := 0; i != b.N; i++ {
+		h.ServeHTTP(w, r)
+	}
+}
+
+type mockResponseWriter struct{}
+
+func (w *mockResponseWriter) WriteHeader(status int)      {}
+func (w *mockResponseWriter) Write(b []byte) (int, error) { return len(b), nil }
+func (w *mockResponseWriter) Header() http.Header         { return http.Header{} }
 
 type mockAddr struct {
 	s string