I noticed logging an HTTP request was quite expensive in both time and memory, so I spent a bit of time to add a few optimizations.

Because some of the optimizations rely on unsafe assumptions about the memory layout of interfaces I made those opt-in, the program may enable them by setting `events.EnableUnsafeOptimizations = true`.

Here are the comparison of running the benchmarks before and after:
```
benchmark                 old ns/op     new ns/op     delta
BenchmarkRequestLog-4     1315          796           -39.47%
BenchmarkHandler-4        1020          918           -10.00%

benchmark                 old allocs     new allocs     delta
BenchmarkRequestLog-4     11             0              -100.00%
BenchmarkHandler-4        1              0              -100.00%

benchmark                 old bytes     new bytes     delta
BenchmarkRequestLog-4     296           0             -100.00%
BenchmarkHandler-4        48            0             -100.00%
```

Please take a look and let me know if something should be changed.