@f2prateek 2017-07-03
From an API standpoint, I don't 100% understand the negative implications of enabling this flag, and I think most people will not as well. So I'm worried people will enable this without understanding the consequences.