@achille-roussel 2017-07-03
OK I'll improve the documentation.