@achille-roussel 2017-07-04
@f2prateek I thought more about your comment and I agree that giving the developer the ability to control when to turn on/off the optimizations is asking for trouble, unless you have a deep understanding of the memory layout used by the compiler you're likely to do the wrong thing. On top of that, it's something you have to do in every program and therefore has a very high maintenance cost to get right.

That being said, for a given version of Go we can for sure know what kind of code the compiler is generating. I changed the approach to use conditional compilation instead, so the developer doesn't need to get involved, and if the compiler ever make changes to how interfaces are represented in memory we can always either disable the optimization for that version of Go, or rewrite them to be compatible with the new version and use the same conditional compilation approach to select the code that gets generated.

I kind of like this approach and it gave me an idea for another optimization that prevents the dynamic memory allocations that are generated for each argument passed to the variable argument list of `Debug(string, ...interface{})` and `Log(string, ...interface{})`.
We now have code which is fully free of dynamic memory allocations caused by generated calls to `runtime.convT2E` (@yields enjoy!)
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkLogger/inline-4     432           350           -18.98%

benchmark                    old allocs     new allocs     delta
BenchmarkLogger/inline-4     3              0              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkLogger/inline-4     24            0             -100.00%
```

Please take a look again and let me know if something should be changed.