@achille-roussel 2017-07-05
@f2prateek 
@yields 
@abraithwaite 

Is there anything more to add or remove in this PR?